Update Language Switcher function
